test(e2e): connect with Google before sign-in
In order to be able to sign in with Google, the user must first connect.
After sign-in, delete Google user account from db.
