+++ tests/e2e/ident-uat-001-connect-with-google.ts
@@ -1,34 +1,8 @@
-import * as admin from 'firebase-admin';
-
 import { NightWatchBrowser } from 'nightwatch';
-
-const serviceAccount = require(process.env.FIREBASE_ADMINSDK_JSON);
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-  databaseURL: process.env.FIREBASE_DATABASE_URL,
-});
+import { deleteGoogleUser } from '../e2e-common';
 
 export = {
-  afterEach: function (_: any, done: Function) {
-    (admin.auth() as any).getUserByEmail(process.env.GOOGLE_TEST_EMAIL)
-      .then((userRecord: any) => {
-        const uid: string = userRecord.uid;
-
-        (admin.auth() as any).deleteUser(uid)
-          .then(() => {
-            console.log(`Deleted ${process.env.GOOGLE_TEST_EMAIL} from database`);
-            done();
-          })
-          .catch((error: Error) => {
-            throw error;
-          });
-      })
-      .catch((error: Error) => {
-        throw error;
-      });
-
-    done();
-  },
+  afterEach: deleteGoogleUser,
 
   'IDENT UAT 1: Connect with Google': function (browser: NightWatchBrowser) {
     browser
